# BackAPI/src/routes/Information.ts
import { Router } from "express";
import prisma from "../database.js";
import { AdminAuthMiddleware } from "../middlewares/AuthMiddleware.js";

const router = Router();

router.get("/", async (req, res) => {
  const { titre, categorie, page, limit } = req.query;
  const informations = await prisma.information.findMany({
    where: {
      titre: titre
        ? { contains: String(titre), mode: "insensitive" }
        : undefined,
      categorie: categorie
        ? { contains: String(categorie), mode: "insensitive" }
        : undefined,
    },
    skip: (Number(page) - 1) * Number(limit),
    take: Number(limit),
    orderBy: { dateModification: "desc" },
  });
  res.json(informations);
});

router.get("/count", AdminAuthMiddleware, async (req, res) => {
  const count = await prisma.information.count();
  res.json({ count });
});

router.get("/categories", async (req, res) => {
  const categories = await prisma.information
    .findMany({
      select: {
        categorie: true,
        status: true,
      },
      distinct: ["categorie"],
    })
    .then((categories) => categories.map((category) => category.categorie));
  res.json(categories);
});

router.post("/create", AdminAuthMiddleware, async (req, res) => {
  const { titre, description, texte, image, categorie, status, authorId } =
    req.body;
  const information = await prisma.information.create({
    data: {
      titre,
      description,
      texte,
      image,
      categorie,
      status,
      authorId,
    },
  });
  res.json(information);
});

router.get("/:id", async (req, res) => {
  const { id } = req.params;
  const information = await prisma.information.findUnique({
    where: {
      id: Number(id),
    },
  });
  res.json(information);
});

router.get("/categorie/:categorie", async (req, res) => {
  const { categorie } = req.params;
  const informations = await prisma.information.findMany({
    where: {
      categorie: String(categorie),
      status: true,
    },
  });
  res.json(informations);
});

router.put("/update/:id", AdminAuthMiddleware, async (req, res) => {
  const { id } = req.params;
  const { titre, description, texte, image, categorie, status, authorId } =
    req.body;
  const information = await prisma.information.update({
    where: {
      id: Number(id),
    },
    data: {
      titre,
      description,
      texte,
      image,
      categorie,
      status,
      authorId,
    },
  });
  res.json(information);
});

router.delete("/delete/:id", AdminAuthMiddleware, async (req, res) => {
  const { id } = req.params;
  await prisma.information.delete({
    where: {
      id: Number(id),
    },
  });
  res.json({ message: "Information deleted successfully" });
});

export default router;